+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - Add `Driver::is_encrypted()`.
  - Reduce the number of lifetime generic parameters in `TransactionQueryBuilder` and `TransactionRecordStream`.
  - Fix `Transaction::rolblack()` failing if a result stream failed before.
+ - Introduce `neo4j::driver::Conifg::with_keep_alive()` and `without_keep_alive()`.
 
 ## 0.0.2
  - Update dependencies.